ETL process for Load Node
This is the cypher script which will load the original data for the load node. In addition, it will load in the newer file, check for any updates to the reference set (PSSE), and update the relationships in the graph. There are 4 different levels of confidence to match, as the source changes ids and sometimes names.

1. 100% confident if the name, psse bus id and id for the load node are all the same
2. 90% confident if the name and id for load node are all the same
3. 80% confident of the busid and load id are the same
4. 60% if memo is the same

For the 90% and 100% it will recreate the relationship to the master node, and use the update as the "most recent reference set". For the lower confidence levels we would need a human to check.

Additionally if no match is found it will create the new master load nodes, and create a relationship to the PSSE file.