Use semi-automated mapping to improve network
as a user defines the node to be an update of, we will keep the confidence at the set "automatically found" level, but add in a property "Manual Override".  This property will cause the graph to update the relationship from IS_CANDIDATE_OF to IS_UPDATE_OF as well as update the relationship between the master node, using the newer node as the reference set.