Update LoadNetwork
there was a bug in creating the relationship of candidates. some of the 80% likely also had a 60% confidence. Adjusted it not to create a relationship if one exists. also added in the ability to automatically map to 80% of the SEER units that are recognized. Mapping of PSSE Generator to Genscape Generator needs to be manual.